[wgsl] Move decl,ptr_spelling to types,pointer.
The pointer validation tests ended up in both decl,ptr_spelling and
types,pointer. This CL moves them into types,pointer to live with the
rest of the type tests.
