+++ src/webgpu/listing_meta.json
@@ -1900,13 +1900,6 @@
   "webgpu:shader,validation,decl,override:no_direct_recursion:*": { "subcaseMS": 1.000 },
   "webgpu:shader,validation,decl,override:no_indirect_recursion:*": { "subcaseMS": 0.951 },
   "webgpu:shader,validation,decl,override:type:*": { "subcaseMS": 12.518 },
-  "webgpu:shader,validation,decl,ptr_spelling:let_ptr_explicit_type_matches_var:*": { "subcaseMS": 1.500 },
-  "webgpu:shader,validation,decl,ptr_spelling:let_ptr_reads:*": { "subcaseMS": 1.216 },
-  "webgpu:shader,validation,decl,ptr_spelling:let_ptr_writes:*": { "subcaseMS": 1.250 },
-  "webgpu:shader,validation,decl,ptr_spelling:ptr_address_space_never_uses_access_mode:*": { "subcaseMS": 1.141 },
-  "webgpu:shader,validation,decl,ptr_spelling:ptr_bad_store_type:*": { "subcaseMS": 0.967 },
-  "webgpu:shader,validation,decl,ptr_spelling:ptr_handle_space_invalid:*": { "subcaseMS": 1.000 },
-  "webgpu:shader,validation,decl,ptr_spelling:ptr_not_instantiable:*": { "subcaseMS": 1.310 },
   "webgpu:shader,validation,decl,var:address_space_access_mode:*": { "subcaseMS": 14.399 },
   "webgpu:shader,validation,decl,var:binding_collision_unused_helper:*": { "subcaseMS": 1.000 },
   "webgpu:shader,validation,decl,var:binding_collisions:*": { "subcaseMS": 1.000 },
@@ -2668,7 +2661,14 @@
   "webgpu:shader,validation,types,matrix:valid:*": { "subcaseMS": 67.784 },
   "webgpu:shader,validation,types,pointer:access_mode:*": { "subcaseMS": 16.078 },
   "webgpu:shader,validation,types,pointer:address_space:*": { "subcaseMS": 523.598 },
+  "webgpu:shader,validation,types,pointer:let_ptr_explicit_type_matches_var:*": { "subcaseMS": 42.865 },
+  "webgpu:shader,validation,types,pointer:let_ptr_reads:*": { "subcaseMS": 38.224 },
+  "webgpu:shader,validation,types,pointer:let_ptr_writes:*": { "subcaseMS": 40.654 },
   "webgpu:shader,validation,types,pointer:missing_type:*": { "subcaseMS": 5.527 },
+  "webgpu:shader,validation,types,pointer:ptr_address_space_never_uses_access_mode:*": { "subcaseMS": 15.312 },
+  "webgpu:shader,validation,types,pointer:ptr_bad_store_type:*": { "subcaseMS": 3.276 },
+  "webgpu:shader,validation,types,pointer:ptr_handle_space_invalid:*": { "subcaseMS": 1.223 },
+  "webgpu:shader,validation,types,pointer:ptr_not_instantiable:*": { "subcaseMS": 12.053 },
   "webgpu:shader,validation,types,pointer:type:*": { "subcaseMS": 11.764 },
   "webgpu:shader,validation,types,ref:not_typeable_alias:*": { "subcaseMS": 6.872 },
   "webgpu:shader,validation,types,ref:not_typeable_let:*": { "subcaseMS": 6.670 },
